Use cpuinfo also for cpu cache size detection.
Encapsulate cache sizes in a CpuCacheSizes struct: this reduces the number of int params passed around block_map.h functions, and provides a single place where to document their meaning. Keep a cpu_cache_sizes.h header just for this struct, allowing BlockMap to depend only on that and not on cpuinfo: that makes it clear that it's only consuming the already-queried cache sizes.

Also add copts=ruy_copts() to many unit tests where it was missing: that ensures in particular that the newly added -Wundef is used in unit tests.

PiperOrigin-RevId: 313639675